fix: passthrough ports
