--- pkg/core/util.go
@@ -1,22 +1,0 @@
-//go:build linux
-
-package core
-
-import (
-	"context"
-
-	"go.keploy.io/server/v2/config"
-)
-
-func GetPortToSendToKernel(_ context.Context, rules []config.BypassRule) []uint {
-	// if the rule only contains port, then it should be sent to kernel
-	ports := []uint{}
-	for _, rule := range rules {
-		if rule.Host == "" && rule.Path == "" {
-			if rule.Port != 0 {
-				ports = append(ports, rule.Port)
-			}
-		}
-	}
-	return ports
-}
